The Krippendorff Module

Currently, there are two computational backends: If the number of possible responses is bounded, one can construct a coincidence matrix with all observed pairs. Computing the observed disagreement is straightforward in this case, while the expected disagreement can easily be cpmputed from the marginals of the coincidence matrix. For all other cases, a generical computation strategy was implemented which avoids constructing temporary tables entirely, but typically scales much worse.

Detailed API

Krippendorff.krippendorffs_alpha — Function

Krippendorff.alpha(...)
krippendorffs_alpha(input; units = rows, metric = :nominal, R = :discrete, silent = false)

Compute the Krippendorff's-α inter-rater reliability measure from the supplied input. The input will be checked to determine how to iterate over it, missing values will be handled automatically if present and the most efficient algorithm to compute α will be determined heuristically. By default, Tables.jl tables and table-like inputs are assumed to have columns representing raters and rows representing units. See prepare_iterator for more information about the input requirements.

Arguments

units::Union{Symbol,AbstractString}: rows or col(umn)s, see prepare_iterator for explanation

metric: a metric computing the squared distance between any pair of responses. Any of [:nominal, :interval] or a custom function. Should satisfy f(x,y) = [0 if x==y], [>0 otherwise] but this is not enforced. See README for explanation of the default metrics.

R: The space of possible responses. Either :discrete (relatively few possible responses, uses fast computation via coincidence matrix), :continuous (many possible responses up to continuous range, uses a slower but generically applicable algorithm with minimal allocation) or a precomputed value (implies discrete, but avoids searching for all unique values). If precomputed, it should be supplied as a tuple or vector of possible values (e.g. the output of unique(yourdata)), or as an appropriate range object where possible (slightly more efficient).

silent::Bool: set to disable all optional output (@info and @warn, doesn't affect error messages)

See also: compute_alpha_generical, compute_alpha_with_coincidences

Missing docstring.

Missing docstring for compute_alpha_with_coincidences. Check Documenter's build log for details.

Missing docstring.

Missing docstring for compute_alpha_generical. Check Documenter's build log for details.

Missing docstring.

Missing docstring for prepare_iterator. Check Documenter's build log for details.

Missing docstring.

Missing docstring for istable. Check Documenter's build log for details.